Adapt SbtScalacParser to SBT 1.x warnings format
The newly released version uses a slightly different warnings format, it prints
the line and column information separated by a colon after the file name.

This leads to an error when clicking on an entry in the warnings list currently,
since the line number is parsed as part of the file name:

```
01 Copying the source file '/local/jenkins/workspace/project/app/actors/SlackActor.scala:25' from the workspace to the build folder '35705b61.tmp' on the Jenkins master failed.
02 Is the file '/local/jenkins/workspace/eve-publish-check/app/actors/SlackActor.scala:25' a valid filename?
```

This commit fixes the regex such that the file name stops at the first colon -
when there is a column number in the warning message.